Address mesh IO #212
