+++ SimPEG/Mesh/TensorMesh.py
@@ -3,6 +3,7 @@ from BaseMesh import BaseMesh, BaseRectangularMesh
 from View import TensorView
 from DiffOperators import DiffOperators
 from InnerProducts import InnerProducts
+from MeshIO import TensorMeshIO
 
 class BaseTensorMesh(BaseMesh):
 
@@ -359,7 +360,7 @@ class BaseTensorMesh(BaseMesh):
 
 
 
-class TensorMesh(BaseTensorMesh, BaseRectangularMesh, TensorView, DiffOperators, InnerProducts):
+class TensorMesh(BaseTensorMesh, BaseRectangularMesh, TensorView, DiffOperators, InnerProducts, TensorMeshIO):
     """
     TensorMesh is a mesh class that deals with tensor product meshes.
